Change datatype for qty field.

+++ MismeAPI/MismeAPI.Common/DTO/Response/EatCompoundDishResponse.cs
@@ -5,6 +5,6 @@ namespace MismeAPI.Common.DTO.Response
     public class EatCompoundDishResponse
     {
         public CompoundDishResponse CompoundDish { get; set; }
-        public int Qty { get; set; }
+        public double Qty { get; set; }
     }
 }
